Zhang community

+++ src/main/java/zhang/community/community/controller/QuestionController.java
@@ -17,7 +17,7 @@ public class QuestionController {
 
     @GetMapping("/question/{id}")
     public String question(
-            @PathVariable(name= "id")Integer id, Model model
+            @PathVariable(name= "id")Long id, Model model
             ){
         QuestionDTO questionDTO=questionService.getById(id);
         //累加阅读数
